+++ admin/client/components/EditForm.js
@@ -180,7 +180,7 @@ var EditForm = React.createClass({
         props.key = field.path;
         // TODO: could we move sortablerelationship to field.type
         if (props.key === 'relateds') {
-          return React.createElement(Fields.sortablerelationship, props);
+          return React.createElement(Fields.sortablerelateds, props);
         } else {
           return React.createElement(Fields[field.type], props);
         }
